Handle that the comment marker can contain regex
In the case of e.g. C Double-tap will correctly detect * as a potential
character a comment can start with. But * has special meaning in regular
expressions and in the case of Vim will trigger an error if not used
correctly.

Protect what Double-tap find by surrounding it in \V...\m. Thanks to
@utrack for the idea.

Fixes #5.